+++ quadra/tasks/anomaly.py
@@ -129,6 +129,7 @@ class AnomalibDetection(Generic[AnomalyDataModuleT], LightningTask[AnomalyDataMo
 
         model_json["image_threshold"] = np.round(self.module.image_threshold.value.item(), 3)
         model_json["pixel_threshold"] = np.round(self.module.pixel_threshold.value.item(), 3)
+        model_json["anomaly_method"] = self.config.model.model.name
 
         with open(os.path.join(self.export_folder, "model.json"), "w") as f:
             json.dump(model_json, f)
@@ -200,10 +201,12 @@ class AnomalibDetection(Generic[AnomalyDataModuleT], LightningTask[AnomalyDataMo
             if any(
                 isinstance(x, MinMaxNormalizationCallback) for x in self.trainer.callbacks  # type: ignore[attr-defined]
             )
-            else self.module.image_metrics.F1Score.threshold
+            else self.module.image_metrics.F1Score.threshold  # type: ignore[union-attr]
         )
 
-        plot_cumulative_histogram(good_scores, defect_scores, threshold.item(), self.report_path)
+        plot_cumulative_histogram(
+            good_scores, defect_scores, threshold.item(), self.report_path  # type: ignore[arg-type, operator]
+        )
 
         _, pd_cm, _ = get_results(np.array(gt_labels), np.array(pred_labels), idx_to_class)
         np_cm = np.array(pd_cm)
@@ -313,7 +316,7 @@ class AnomalibEvaluation(Evaluation[AnomalyDataModule]):
         self.datamodule.setup(stage="test")
         test_dataloader = self.datamodule.test_dataloader()
 
-        optimal_f1 = OptimalF1(num_classes=None, pos_label=1)
+        optimal_f1 = OptimalF1(num_classes=None, pos_label=1)  # type: ignore[arg-type]
 
         anomaly_scores = []
         anomaly_maps = []
@@ -326,7 +329,11 @@ class AnomalibEvaluation(Evaluation[AnomalyDataModule]):
                 batch_labels = batch_item["label"]
                 image_labels.extend(batch_labels.tolist())
                 image_paths.extend(batch_item["image_path"])
-                anomaly_map, anomaly_score = self.deployment_model(batch_images.to(self.device))
+                if self.model_data.get("anomaly_method") == "efficientad":
+                    model_output = self.deployment_model(batch_images.to(self.device), None)
+                else:
+                    model_output = self.deployment_model(batch_images.to(self.device))
+                anomaly_map, anomaly_score = model_output[0], model_output[1]
                 anomaly_map = anomaly_map.cpu()
                 anomaly_score = anomaly_score.cpu()
                 known_labels = torch.where(batch_labels != -1)[0]