+++ CHANGELOG.md
@@ -11,6 +11,11 @@ All notable changes to this project will be documented in this file.
 - Add support for onnx model import in all evaluation tasks.
 - Add `export` configuration group to regulate exportation parameters.
 - Add `inference` configuration group to regulate inference parameters.
+- Add EfficientAD configuration for anomaly detection.
+
+#### Updated
+
+- Update anomalib library from version 0.4.0 to 0.7.0
 
 #### Changed 
 
@@ -21,6 +26,7 @@ All notable changes to this project will be documented in this file.
 #### Fixed
 
 - Fix failures when trying to override `hydra` configuration groups due to wrong override order.
+- Fix certain anomalib models not loaded on the correct device.
 
 ### [1.1.4]